Refactor Bluetooth handling and UI
